requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.0,<5.0
django-filter>=23.0,<24.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
drf-spectacular>=0.27,<1.0
reportlab>=4.0,<5.0
cloudinary>=1.36,<2.0
whitenoise>=6.5,<7.0
django-ratelimit>=4.1,<5.0
